backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.7.2

# ML/AI
scikit-learn>=1.3.0
tensorflow>=2.15.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Data Processing
joblib==1.3.2

# SSH/Remote Execution
paramiko==3.4.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Reporting
reportlab==4.0.7
weasyprint==60.1
python-docx==1.1.0
markdown2==2.4.10
Jinja2==3.1.2
pygments==2.17.2

# Charts and visualization
matplotlib==3.8.2
seaborn==0.13.0

# Additional ML for agent intelligence
sentence-transformers==2.2.2
langchain==0.1.0

# PDF manipulation
PyPDF2==3.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Tools Module Dependencies
aiohttp>=3.8.0
threadpoolctl>=3.0.0
